Added initial guide and removed sto_lim suffix
