--- mpl_style.py
@@ -1,176 +1,0 @@
-# -*- coding: utf-8 -*-
-"""
-Created on Tue Nov  3 14:14:09 2020
-
-@author: user
-"""
-#box style
-style1 = {
-    # Line styles
-    'lines.linewidth': 1.5,
-    'lines.antialiased': True,
-
-    # Font
-    'font.size': 16.0,
-
-    # Axes
-    'axes.linewidth': 1.5,
-    'axes.titlesize': 'x-large',
-    'axes.labelsize': 'large',
-    'axes.spines.top': True,
-    'axes.spines.right': True,
-
-    # Ticks
-    'xtick.major.size': 6,
-    'xtick.minor.size': 4,
-    'xtick.major.width': 1.5,
-    'xtick.minor.width': 1.5,
-    'xtick.major.pad': 6,
-    'xtick.minor.pad': 6,
-    'xtick.labelsize': 'medium',
-    'xtick.direction': 'in',
-    'xtick.top': False,
-
-    'ytick.major.size': 6,
-    'ytick.minor.size': 4,
-    'ytick.major.width': 1.5,
-    'ytick.minor.width': 1.5,
-    'ytick.major.pad': 6,
-    'ytick.minor.pad': 6,
-    'ytick.labelsize': 'medium',
-    'ytick.direction': 'in',
-    'ytick.right': False,
-
-    # Legend
-    'legend.fancybox': True,
-    'legend.fontsize': 'large',
-    'legend.scatterpoints': 5,
-    'legend.loc': 'best',
-
-    # Figure
-    'figure.figsize': [8, 6],
-    'figure.titlesize': 'large',
-
-    # Images
-    'image.cmap': 'magma',
-    'image.origin': 'lower',
-
-    # Saving
-    'savefig.bbox': 'tight',
-    'savefig.format': 'png',
-}
-
-#Just y and x axis
-style2 = {
-    # Line styles
-    'lines.linewidth': 1.5,
-    'lines.antialiased': True,
-
-    # Font
-    'font.size': 16.0,
-
-    # Axes
-    'axes.linewidth': 1.5,
-    'axes.titlesize': 'x-large',
-    'axes.labelsize': 'large',
-    'axes.spines.top': False,
-    'axes.spines.right': False,
-
-    # Ticks
-    'xtick.major.size': 6,
-    'xtick.minor.size': 4,
-    'xtick.major.width': 1.5,
-    'xtick.minor.width': 1.5,
-    'xtick.major.pad': 6,
-    'xtick.minor.pad': 6,
-    'xtick.labelsize': 'medium',
-    'xtick.direction': 'in',
-    'xtick.top': False,
-
-    'ytick.major.size': 6,
-    'ytick.minor.size': 4,
-    'ytick.major.width': 1.5,
-    'ytick.minor.width': 1.5,
-    'ytick.major.pad': 6,
-    'ytick.minor.pad': 6,
-    'ytick.labelsize': 'medium',
-    'ytick.direction': 'in',
-    'ytick.right': False,
-
-    # Legend
-    'legend.fancybox': True,
-    'legend.fontsize': 'large',
-    'legend.scatterpoints': 5,
-    'legend.loc': 'best',
-
-    # Figure
-    'figure.figsize': [8, 6],
-    'figure.titlesize': 'large',
-
-    # Images
-    'image.cmap': 'magma',
-    'image.origin': 'lower',
-
-    # Saving
-    'savefig.bbox': 'tight',
-    'savefig.format': 'png',
-}
-
-
-#Just y and x axis, no ticks
-style3 = {
-    # Line styles
-    'lines.linewidth': 1.5,
-    'lines.antialiased': True,
-
-    # Font
-    'font.size': 16.0,
-
-    # Axes
-    'axes.linewidth': 1.5,
-    'axes.titlesize': 'x-large',
-    'axes.labelsize': 'large',
-    'axes.spines.top': False,
-    'axes.spines.right': False,
-
-    # Ticks
-    'xtick.major.size': 0,
-    'xtick.minor.size': 0,
-    'xtick.major.width': 0,
-    'xtick.minor.width': 0,
-    'xtick.major.pad': 0,
-    'xtick.minor.pad': 0,
-    'xtick.top': False,
-    'xtick.bottom': False,
-
-    'ytick.major.size': 0,
-    'ytick.minor.size': 0,
-    'ytick.major.width': 0,
-    'ytick.minor.width': 0,
-    'ytick.major.pad': 0,
-    'ytick.minor.pad': 0,
-    'ytick.right': False,
-    'ytick.left': False,
-    
-    'xtick.labelsize' : 16,
-    'ytick.labelsize' : 16,
-    
-
-    # Legend
-    'legend.fancybox': True,
-    'legend.fontsize': 'large',
-    'legend.scatterpoints': 5,
-    'legend.loc': 'best',
-
-    # Figure
-    'figure.figsize': [8, 5],
-    'figure.titlesize': 'large',
-
-    # Images
-    'image.cmap': 'magma',
-    'image.origin': 'lower',
-
-    # Saving
-    'savefig.bbox': 'tight',
-    'savefig.format': 'png',
-}
