+++ pyfpt/analytics/__init__.py
@@ -1,11 +1,12 @@
 from .classicality_criterion import classicality_criterion
-from .edgeworth_pdf_sto_limit import edgeworth_pdf_sto_limit
-from .fourth_central_moment_N_sto_limit import\
-    fourth_central_moment_N_sto_limit
-from .gaussian_deviation_sto_limit import\
-    gaussian_deviation_sto_limit
-from .kurtosis_N_sto_limit import kurtosis_N_sto_limit
-from .mean_N_sto_limit import mean_N_sto_limit
+from .edgeworth_pdf import edgeworth_pdf
+from .gaussian_pdf import gaussian_pdf
+from .fourth_central_moment_N import\
+    fourth_central_moment_N
+from .gaussian_deviation import\
+    gaussian_deviation
+from .kurtosis_N import kurtosis_N
+from .mean_N import mean_N
 from .quadratic_inflation_large_mass_pdf import\
     quadratic_inflation_large_mass_pdf
 from .quadratic_inflation_near_tail_pdf import\
@@ -13,7 +14,7 @@ from .quadratic_inflation_near_tail_pdf import\
 from .reduced_potential import reduced_potential
 from .reduced_potential_ddiff import reduced_potential_ddiff
 from .reduced_potential_diff import reduced_potential_diff
-from .skewness_N_sto_limit import skewness_N_sto_limit
-from .third_central_moment_N_sto_limit import\
-    third_central_moment_N_sto_limit
-from .variance_N_sto_limit import variance_N_sto_limit
+from .skewness_N import skewness_N
+from .third_central_moment_N import\
+    third_central_moment_N
+from .variance_N import variance_N
