+++ setup.cfg
@@ -1,6 +1,6 @@
 [metadata]
 name = PyFPT
-version = 0.0.2
+version = 0.0.5
 author = Joe Jackson
 author_email = joseph.jackson@port.ac.uk
 description = This is package for calculating the first-passage times of stochastic simulations.
@@ -25,6 +25,7 @@ setup_requires =
 packages = find:
 install_requires =
 	matplotlib
+	mpmath
 	numpy
 	scipy
 	pandas
